Updated Transform and Movement Components with Vectors
-Replaced xPixel and yPixel fields with a Vector3D position
-Deleted the RotationComponent and added a rotaion field to
TransformComponent
-Update all fields in CameraComponent to Floats
-Replaced MovementComponent with a PhysicsComponent containing a
Vector3D containing forces acting on it.
-Updated all systems apart from Collision and Movement to use the new
Vector based components.
-Removed BattleGameState, CollisionSystem, and MovementSystem from the
buildpath. To be updated in the next revision.

+++ src/aaron/game/pokemonatb/component/CameraComponent.java
@@ -2,12 +2,12 @@ package aaron.game.pokemonatb.component;
 
 public class CameraComponent extends Component{
 	//in pixels top left hand pixel
-	public int xPos;
-	public int yPos;
-	public int xSize;
-	public int ySize;
+	public float xPos;
+	public float yPos;
+	public float xSize;
+	public float ySize;
 	
-	public CameraComponent(int xPos, int yPos, int xSize, int ySize) {
+	public CameraComponent(float xPos, float yPos, float xSize, float ySize) {
 		super();
 		this.xPos = xPos;
 		this.yPos = yPos;
